+++ src/App.jsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 
-import Navbar from './components/Navbar';
+import Navbar from './components/navbar';
 
 export default class App extends Component {
   render() {
